🔧 Enhancement: API Extensions and Real Server Integration for Advanced Features
- Backend: Add new endpoints for challenge instruments, portfolio, and orders
- Backend: Implement GetChallengeInstrumentsService, GetPortfolioService, GetOrdersService
- Backend: Add corresponding DTOs and fix type casting bug in GetChallengesService
- Frontend: Integrate real API calls in portfolio optimization page with React Query
- Frontend: Add loading states, error handling, and mutation functions
- Testing: Enhance E2E tests with real server communication validation
- Testing: Add API call monitoring and server health checks

+++ backend/src/main/java/com/stockquest/application/challenge/GetChallengesService.java
@@ -66,7 +66,7 @@ public class GetChallengesService {
      */
     @Cacheable(value = "activeChallengeCount", cacheManager = "cacheManager")
     public long getActiveChallengeCount() {
-        // 임시로 featured 챌린지 수를 반환
-        return challengePort.findFeaturedChallenges().size();
+        // 임시로 featured 챌린지 수를 반환 (size()는 int이므로 long으로 명시적 캐스팅)
+        return (long) challengePort.findFeaturedChallenges().size();
     }
 }
